% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN_add_missing_data_for_grouping.R
\name{add_missing_data_for_grouping}
\alias{add_missing_data_for_grouping}
\title{Add missing data for a single grouping}
\usage{
add_missing_data_for_grouping(data)
}
\arguments{
\item{data}{A tibble containing the data to be used for time series
forecasting, which has been created using the
\code{initialize_ts_forecast_data} function.}
}
\value{
A tibble which has been extended with previously missing data points
  for a single grouping.
}
\description{
\code{add_missing_data_for_grouping_forecast_package} is a function to add missing data points
within the range of dates available in a dataset to be used for time series
forecasting, for only a single available grouping. The function creates
additional dummy data for historic data that is missing. It will impute the
missing months with a specified (non-zero) value, so that the time series
forecasting models can handle the data.
}
\examples{
initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   dplyr::mutate(
      col_of_interest = ifelse(
         test = period \%in\% as.Date(c("1995-01-31", "1996-03-31", "2000-08-31")),
         yes = NA,
         no = col_of_interest
      )
   ) \%>\%
   add_missing_data_for_grouping()
}
