% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_add_up_hierarchical_values.R
\name{add_up_hierarchical_values}
\alias{add_up_hierarchical_values}
\title{Add up hierarchical values}
\usage{
add_up_hierarchical_values(data, hierarchy)
}
\arguments{
\item{data}{A tibble containing the data to be used for time series
forecasting, which has been created using the
\code{initialize_ts_forecast_data} function and which is hierarchical.}

\item{hierarchy}{A list containing the hierarchy matrix and data showing the
levels for each hierarchical group. It is assumed to be created using the
\code{create_hierarchy_list} function.}
}
\value{
A tibble which is similar to input data, except that all the parent
  hierarchical groups are re-created by adding up their respective
  hierarchical children.
}
\description{
\code{add_up_hierarchical_values} is a function to make the hierarchical
dataset consistent within itself.
}
\examples{
data <- dummy_hierarchical_gasprice \%>\%
   initialize_ts_forecast_data(
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = "currency",
      hierarchical_cols = c("location", "oil_company")
   ) \%>\%
   dplyr::mutate(original_col_of_interest = col_of_interest)
hierarchy <- create_hierarchy_list(data)
add_up_hierarchical_values(
   data = data,
   hierarchy = hierarchy
)
}
