% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CLEAN_detect_anomalies_using_local_outlier_factor.R
\name{detect_anomalies_using_local_outlier_factor}
\alias{detect_anomalies_using_local_outlier_factor}
\title{Detect anomalies using local outlier factor}
\usage{
detect_anomalies_using_local_outlier_factor(
  data,
  threshold = 10,
  s = 9,
  m_width = 3,
  per_group = FALSE
)
}
\arguments{
\item{data}{A tibble containing the data to be cleaned before it is used for
time series forecasting, which has been created using the
\code{initialize_ts_forecast_data} function.}

\item{threshold}{A positive numeric value to label datapoints as anomaly or
not. Higher values correspond to a more conservative threshold, making
points less likely to be labeled as anomalies.}

\item{s}{A positive integer value corresponding to the size of the
neighborhood of data points to be considered when running the algorithm.}

\item{m_width}{A positive integer value indicating the number of data point
to use when calculating the moving average and moving standard deviation
that are used to flag outliers.}

\item{per_group}{Boolean, which is set to TRUE to calculate the local outlier
factor separately for each group in the data, or set to FALSE to calculate
the local outlier factor using all per_group in the data together.}
}
\value{
A tibble extended with several columns related to anomaly detection.
}
\description{
\code{detect_anomalies_using_local_outlier_factor} is a function to detect
potential anomalies by using a machine learning technique called local
outlier factor. The technique attempts to find anomalies based on the
absolute percentage difference to the previous time steps moving average, as
well as information on which month and quarter it is. The local outlier
factor gives an anomaly score which is used to determine if a data point is
considered a anomaly using a cut-off point.
}
\examples{
detect_anomalies_using_local_outlier_factor(dummy_anomaly_data)
}
