% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CLEAN_detect_anomalies_using_percent_differences.R
\name{detect_anomalies_using_percent_differences}
\alias{detect_anomalies_using_percent_differences}
\title{Detect anomalies using a percentage difference}
\usage{
detect_anomalies_using_percent_differences(
  data,
  max_pct_diff = 0.3,
  m_width = 3
)
}
\arguments{
\item{data}{A tibble containing the data to be cleaned before it is used for 
time series forecasting, which has been created using the 
\code{initialize_ts_forecast_data} function.}

\item{max_pct_diff}{A scalar between 0 and 1, which defines the maximum 
percentage difference between two consecutive values of the time series for 
it to be considered a potential anomaly.}

\item{m_width}{A positive integer value indicating the number of data point 
to use when calculating the moving average flag outliers.}
}
\value{
A tibble extended with several columns related to anomaly detection.
}
\description{
\code{detect_anomalies_using_percent_differences} is a function to detect
potential anomalies by calculating a percentage difference of every point
compared to the moving average at the previous timestep. Combined with a 
specified cut-off point this will flag individual points as potential anomalies.
}
\examples{
detect_anomalies_using_percent_differences(dummy_anomaly_data)
}
