% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN_detect_anomalies_using_tsoutlier.R
\name{detect_anomalies_using_ts_outlier}
\alias{detect_anomalies_using_ts_outlier}
\title{Detect anomalies using the tsoutlier function}
\usage{
detect_anomalies_using_ts_outlier(data)
}
\arguments{
\item{data}{A tibble containing the data to be cleaned before it is used for
time series forecasting, which has been created using the
\code{initialize_ts_forecast_data} function.}
}
\value{
A tibble extended with several columns related to anomaly detection.
}
\description{
\code{detect_anomalies_using_ts_outlier} is a function to detect potential
anomalies using the tsoutliers function from the forecast package.
}
\examples{
data <- detect_anomalies_using_ts_outlier(dummy_anomaly_data)
}
