% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_get_fc_date.R
\name{get_fc_date}
\alias{get_fc_date}
\title{Convert scenario string into a forecast date}
\usage{
get_fc_date(misob_year, scenario)
}
\arguments{
\item{misob_year}{A four digit integer/numeric value indicating the MISOB
year which was used to retreive the MISOB data.}

\item{scenario}{A character string coming from the scenario column in a MISOB
dataset as retrieved from MISOB.}
}
\value{
An object in yyyymm numeric format
}
\description{
\code{get_fc_date} is a function to convert a character string containing
information on the scenario (as retrieved from MISOB using the
\code{retrieve_misob_data} function) into a period (numeric - yyyymm format).
}
\examples{
get_fc_date(
   misob_year = 2018,
   scenario = "Forecast May"
)
}
