% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN_load_saved_anomalies.R
\name{load_saved_anomalies}
\alias{load_saved_anomalies}
\title{Loads saved anomalies from a csv file from a file_path.}
\usage{
load_saved_anomalies(file_path)
}
\arguments{
\item{file_path}{The file_path to the csv file with the anomalies.}
}
\value{
A tibble containing the anomalies
}
\description{
\code{load_saved_anomalies} is a function that returns saved anomalies from a
csv file.
}
\examples{
# REPLACE network_location_path/file_name WITH A VALID FILE ON A NETWORK LOCATION
load_saved_anomalies(
  file_path = "network_location_path/file_name.csv"
)
}
