% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VISUALIZE_plot_anomalies.R
\name{plot_anomalies}
\alias{plot_anomalies}
\title{Plot anomalies for visual inspection}
\usage{
plot_anomalies(data)
}
\arguments{
\item{data}{A tibble containing the data on which anomaly detection has been
run, containing a single grouping or multiple.}
}
\value{
A plotly object highlighting points as potential anomalies. If the data 
  multiple groupings. It will return a facet plot with simpler features.
}
\description{
\code{plot_anomalies} is a function to create a plot which displays the
development of data over time, in which potential anomalies that have been
detected are highlighted for visual inspection.
}
\examples{
initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company")
   ) \%>\%
   detect_anomalies_using_moving_standard_scores() \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   plot_anomalies()
}
