% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN_remove_saved_anomalies.R
\name{remove_saved_anomalies}
\alias{remove_saved_anomalies}
\title{Removes anomalies from data frame}
\usage{
remove_saved_anomalies(data, anomalies)
}
\arguments{
\item{data}{A tibble containing the data to be cleaned before it is used for
time series forecasting, which has been created using the
\code{initialize_ts_forecast_data} function.}

\item{anomalies}{A tibble containing the data with the anomalies. Can be
loaded by using the \code{load_saved_anomalies} function.}
}
\value{
A tibble possibly extended with is_anomaly column and anomality rows
  imputed with NaN.
}
\description{
\code{remove_saved_anomalies} is a function that takes a prepared dataframe
and imputes Nan to rows from a separate dataframe containin the rows with
anomalies. The function can be applied to already cleaned data where the
function will only impute additional rows from the anomalies data frame.
}
