% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRAPPER_start_data_cleaning_dashboard.R
\name{start_data_cleaning_dashboard}
\alias{start_data_cleaning_dashboard}
\title{Start a local data cleaning dashboard}
\usage{
start_data_cleaning_dashboard(data, add_total_per_group = T)
}
\arguments{
\item{data}{A tibble containing several columns of data required for time
series forecasting, which has been created using the
\code{initialize_ts_forecast_data} function.}

\item{add_total_per_group}{Boolean, which is set to TRUE if for every group
in the data (with more than one level) a total should be added that is
shown in the dashboard, or set to FALSE if not.}
}
\value{
Starts a local Shiny dashboard to execute data cleaning.
}
\description{
\code{start_data_cleaning_dashboard} is a wrapper function around several
visualization functions to start a local Shiny dashboard to execute the
cleaning of the available data.
}
\examples{
start_data_cleaning_dashboard(dummy_anomaly_data)
}
